When a commonality references a commonality from a different concept metamodel, the reactions generator fails to lookup the domain for the referenced concept.
This is solved by (temporarily) registering all concept domains during the reactions generation (instead of only registering the domain for the concept of the currently handled CommonalityFile).

The GeneratorContext keeps track of the concept names of the currently processed ResourceSet. The workaround for identifying the last seen ResourceSet was moved from the IntermediateModelGenerator into the GenerationContext. The IntermediateModelCodeGenerator and the DomainGenerator skip execution when the concepts of the ResourceSet have already been processed.

GenerationContext#getResourceSet is accessible in the various sub generators now.
Also added some debug output to the reactions generation.