When using multiple concepts, the shared namespace URI of the 'resources' metamodel leads to conflicts during domain registration.
By default Vitruvius uses an object's package namespace uri to check if an object belongs to a given domain. We removed the 'resources' namespace URI from the intermediate domains and added a special case for the common IntermediateResourceBridge class: Instances of it get an intermediate namespace uri assigned at construction and we use that to determine the intermediate concept domain it belongs to.