Fixed: Commonality participations could not get matched yet because their implicit intermediate model root has not been taken into account previously.
Since the intermediate model root is not explicitly specified in a participation's declaration, we do not have a corresponding ParticipationClass to represent it and its containment relations. Instead of adding a dummy metaclass to explicitly represent the intermediate model root in commonality participations (similar to the Resource metaclass for regular participations), and instead of changing how we represent participation roots, contexts and containments, we allow the commonality participation's own root to be empty now and adapted any code that deals with participation contexts and matching of participations to be able to deal with empty participation roots and commonality participations specifically.

When creating the ContainmentTree for the ParticipationMatcher, we create a node and edges to represent the implicit intermediate model root and its containment relations. The ParticipationMatcher now considers intermediate model roots as root to start the participation matching at.

Instead of assuming that any participation in its own specified context declares a resource class, a resource bridge is only setup for participations that actually declare a resource class (i.e. no resource bridges are setup for commonality participations).

Minor refactoring in ParticipationContextHelper:
* ParticipationContext has access to the ParticipationContextHelper now.
* The ParticipationContext's non-root containments get dynamically calculated now using the helper.
* ParticipationContext#getParticipationClasses has been adapted to be able to deal with empty roots.
* ParticipationRoot#getNonRootBoundaryClasses has been removed in favour of a single general getNonRootBoundaryClasses method inside ParticipationContextHelper.
* ParticipationContextHelper#getNonRootClasses has been simplified.