Operators are now weakly cached by ResourceSet.
This ensures that the cached operators do not persist across Xtext builds.
However (TODO), content assist uses the same ResourceSet while the editor for some commonality file is open, resulting in the cache not getting cleared for content assist as long as the editor is open.

Also added a note about a similar potential caching issues for ParticipationRoots and ParticipationContents.