Fixed: ResourceBridges for singletons require an intermediate correspondence.
Previously we intended to omit this correspondence, since singletons are not bound to a single participation and intermediate instance. However, they use this correspondence to figure out the intermediate model resource they are stored in, as well as the project's source folder within which the corresponding resource gets persisted. We now add a correspondence for the first intermediate for which the singleton is used.

Other related changes:
* The ResourceBridge now uses its own intermediate correspondence instead of relying on the intermediate correspondence of the contained object.
* Added a note regarding the risk of accessing singleton objects from within mappings. This should ideally be enforced via validation at some point.