Fixed: The ModelTreeEqualityMatcher now ignores all derived features.
This resolves an endless recursion issue when following reverse references.

This is not implemented as a FeatureMatcher but enforced by default for all object comparisons, because the comparison of derived features should already be covered by the comparison of the non-derived features from which they derive their value and therefore be safe to be ignored.