Fixed: When using multiple participations classes for the same metaclass, change propagation would sometimes not come to an end due to cyclic changes.
The problem was that the participation attribute change reactions would react to any changes of the affected metaclass, instead of only those involving a specific participation class. We now retrieve the corresponding intermediate using the correspondence tag specific to the participation class involved in the attribute mapping. This way any changes to objects of other participation classes with the same metaclass are ignored.