Implemented reference mapping operators for attribute references.
'Attribute references' occur when a commonality reference is implicitly represented by certain conditions for the attribute values of the involved participation objects. For example, while UML packages have containment references to their sub-packages, in Java this sub-package relationship is represented implicitly by the namespaces and name attributes of the involved Java packages.

Grammar changes: CommonalityReferenceMapping has two subtypes now, one for regular reference mappings and one for the new operator reference mappings. The previous ReferenceEqualitySpecification, ReferenceSetSpecification and ReferenceReadSpecification subtypes have been replaced with boolean read and write attributes. Reference mapping operators can refer to attributes of the referenced participation using the 'ref' keyword.

The participation attribute operands of the reference mapping operator identify the attributes involved in the attribute reference. Whenever one of these attributes changes, we check if the previous references are still valid and if any new references can be matched. Literal operands can be used to pass additional parameters to the operator at runtime.

An attribute change of the container object may result in multiple new attribute containment relations to be established. The matching is therefore performed multiple times, until no more matches are found. There is a new 'matchManyParticipations' routine, which invokes the matching routine as often as new participation matches are found.

The ParticipationRoot, ParticipationContext, ContainmentTree and ParticipationMatcher have been updated to additionally represent one external attribute reference root and its attribute references to objects of the referenced participation. There are now Containment sub-types for regular reference containments and containments implemented by an operator. If the ContainmentTree specifies an attribute reference root, the specified root intermediate type applies to that attribute reference root instead.

Matching participations referenced via attribute reference mappings uses the referenced participation's regular root context and additionally checks if the attribute references are fulfilled. The ParticipationMatcher and the matching routine have a new 'followAttributeReferences' input parameter, which is used when we want to match the referenced participations for a given attribute reference root object. The ParticipationMatcher will ask the attribute reference operator(s) for the objects contained by the attribute reference root object and then continues to follow the containment hierarchies of these objects upwards in order to find candidate root objects for the matching to start at. This may return more than one candidate root object now.

Whether the reference declared by an operator mapping is multi-valued and whether it actually is an attribute reference is specified inside the operator's annotation. If the operator is not marked as 'attribute reference' it is handled like a regular reference mapping, except that the ParticipationMatcher uses the operator to determine the contained objects. However, this is not fully supported yet.

A reference mapping operator provides operations for getting the objects contained by a given object, getting the container for a given object, checking if a given object is contained by another given object, and inserting one object into another (i.e. setting the relevant attribute values accordingly in the case of an attribute reference).

In order to ensure that attribute reference operators have access to the fully initialized attributes of newly instantiated participation objects, we first run all participation object post initializers and a new applyCommonalityAttributes routine applies all intermediate attributes to the freshly created participation objects before the containment relations get established.

Other related changes:
* ParticipationContext has new is/getReferenceRootClass methods which handle the special case of having an attribute reference root. All previous calls to ParticipationContext#getRootContainerClass have been updated to use this method instead.
* Participation contexts may now have two 'roots': The participation's own root and the attribute reference root specified by a reference mapping. In order to avoid confusion, the 'participationContextRoot' variables have been renamed to 'referenceRoot'.
* Since we cannot easily invoke the object deletion action from within execution blocks, we generate a 'deleteObject' routine which simply deletes any passed object.
* Added validation: Reference mapping operators need to specify at least one referenced participation attribute via their operands.
* Added validation: The participations referenced by reference mapping operators need to declare a root context.
* We skip the validation of reference mappings if the retrieved participation is null. This can happen if the mapping is still incomplete at the time of validation.

Other changes:
* Grammar: Added a common type 'Operand' which all types of operands extend from.
* Moved general operand utility methods into OperandExtension.
* Added reactions runtime dependency to manifest.
* Renamed the 'matchCommonalityReferenceMapping' routines to 'matchReferenceMapping'.
* Minor variable name change: The 'insertReferencedIntermediate' routine now refers to the 'referencingIntermediate' instead of just 'intermediate'.