Added attribute mapping operator tests.
The tests cover:
* Converting between single valued primitive attributes.
* Converting between single and multi-valued attributes.

The used operators cannot be located inside the ui tests project, because this would require the runtime test project to have a dependency to the ui tests project, which would then result in all the commonality files of the other ui tests being included inside the build. The operators are therefore located inside the commonalities testutils project.