HierarchicalTuidCalculatorAndResolver no longer restricts the 'root' objects to be of the meta-model's root type.
The cache for not yet persisted root objects inside TuidCalculatorAndResolverBase is able to also contain objects which are not of the meta-model's root type if these objects have not been inserted into any root container yet at the time the TUID for them or one of their contained objects is calculated.
Removing the check for the meta-model's root type does not seem to cause any issues and allows us to also resolve objects within these root objects of non-root type.