Fixed: The ExternalReactionsGenerator would sometimes setup the wrong type of runtime module.
We now query the IGlobalServiceProvider to trigger the setup of the correct runtime module depending on the current context (UI/Eclipse or standalone).
However, depending on whether the client has already started using other parts of the Reactions language (such as the FluentReactionsLanguageBuilder) this setup may be too late. Clients should therefore trigger the setup of the Reactions language themselves earlier already.