Fixed: JaMoPP classifiers for Java library classes cannot be resolved in some cases.
The issue is that when the UuidGeneratorAndResolver tries to resolve such a classifier, JaMoPP will load the corresponding resource and then modify the name of the freshly loaded CompilationUnit. The notification for this model change leads to an exception being thrown if we are not inside a write transaction currently. The UuidGeneratorAndResolver will catch this exception and assume that the resolution has failed.

Retrying the resolution resolves this issue for the most part.