Fixed: Participation attribute names which are in conflict with language keywords could not fallback to using the caret (^) prefix.
The issue is that we are not using the ID rule directly, but delegate to it via our own defined datatype parser rules. This has the effect that the ID value converter (which would usually remove the caret character from the id) does not get invoked for our datatype parser rules.

I simplified the datatype parser rules and added an extended value converter service which delegates to the value converter of the 'SimpleID' parser rule. The grammar also uses SimpleID in place of ID in several places now.

A new 'QualifiedClassValueConverter' handles the delegation for the 'QualifiedClass' datatype parser rule.