Add option to pass IdResolver into change propagation
In order to be able to cross resolve the affected and changed objects from a change the used id resolver has to cross resolve between the view and its source. This is achieved by overloading the relevant methods with an optional idresolver which if present will be used instead of a default one.